+++ colorbleed/plugins/maya/load/load_rig.py
@@ -22,8 +22,6 @@ class RigLoader(api.Loader):
 
     def process(self, name, namespace, context, data):
 
-        assetname = "{}_".format(context["asset"]["name"])
-        unique_namespace = maya.unique_namespace(assetname, format="%03d")
         nodes = cmds.file(self.fname,
                           namespace=namespace,
                           reference=True,
@@ -34,7 +32,7 @@ class RigLoader(api.Loader):
         # Store for post-process
         self[:] = nodes
         if data.get("post_process", True):
-            self._post_process(name, unique_namespace, context, data)
+            self._post_process(name, namespace, context, data)
 
     def _post_process(self, name, namespace, context, data):
         from avalon import maya
