+++ colorbleed/plugins/maya/load/load_setdress_rebuild.py
@@ -4,10 +4,10 @@ from avalon import api
 class SetDressRebuild(api.Loader):
 
     families = ["colorbleed.setdress"]
-    representations = ["abc"]
+    representations = ["json"]
 
     label = "Rebuild Set Dress"
-    order = -10
+    order = -9
     icon = "code-fork"
     color = "orange"
 
@@ -17,12 +17,14 @@ class SetDressRebuild(api.Loader):
         from maya import cmds
         from avalon.tools.cbloader import lib
         from colorbleed.maya import lib as clib
+        from cb.utils.maya import core
 
-        # Ensure
-        data_file = self.fname.replace(".abc", ".json")
-        with open(data_file, "r") as fp:
+        with open(self.fname, "r") as fp:
             build_data = json.load(fp)
 
+        cmds.namespace(add=namespace)
+
+        new_transforms = []
         for representation_id, instances in build_data.items():
 
             # Find the corresponding loader
@@ -30,7 +32,7 @@ class SetDressRebuild(api.Loader):
 
             # Ensure context can be passed on
             for inst in instances:
-                # Get the uses loader
+                # Get the used loader
                 Loader = next((x for x in loaders if
                                x.__name__ == inst['loader']),
                               None)
@@ -41,18 +43,24 @@ class SetDressRebuild(api.Loader):
                     continue
 
                 # Run the loader
-                namespace = inst['namespace'].strip(":")
+                instance_ns = ":".join([namespace,
+                                        inst['namespace'].strip(":")])
                 container = lib.run_loader(Loader,
                                            representation_id,
-                                           namespace=namespace)
-
-                # Apply transformations
-                if not inst["matrix"]:
-                    continue
+                                           namespace=instance_ns)
 
                 container_data = {"objectName": container}
                 transforms = clib.get_container_transforms(container_data)
                 # Force sort order, similar to collector
                 transforms = sorted(transforms)
+
                 for idx, matrix in inst["matrix"].items():
                     cmds.xform(transforms[int(idx)], matrix=matrix)
+
+                # Store the created container
+                self.append(container)
+                new_transforms.extend(transforms)
+
+        roots = core.getHighestInHierarchy(new_transforms)
+        cmds.group(roots, name="{}:{}".format(namespace,
+                                              context['subset']['name']))
